requirements.txt
numpy>=1.21
scipy>=1.7
mne>=1.2
antropy>=0.1.6
joblib>=1.1
scikit-learn>=1.0
hydra-core>=1.3
omegaconf>=2.3
mlflow>=2.0
networkx>=2.6
torch>=1.12
shap>=0.41.0
imbalanced-learn>=0.10
pytest>=7.0
typeguard>=2.13
matplotlib>=3.5
